% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmlParty.R
\name{pmmlparty}
\alias{pmmlparty}
\title{Generate PMML for party objects}
\usage{
pmmlparty(model, formula, data = data, model.name = "CHAID_Model",
  app.name = "Rattle/PMML", description = "CHAID Decision Tree Model",
  copyright = NULL, transforms = NULL, unknownValue = NULL,
  dataset = NULL, ...)
}
\arguments{
\item{model}{an object of class party}

\item{formula}{an object of class formula}

\item{data}{dataset used for training the model}
}
\value{
An object of class pmml
}
\description{
Generate the PMML representation for a party object from package party and partykit
}
\details{
PMML is an XML based language which provides a way for applications to define statistical and data mining models and to share models between PMML compliant applications. More information about PMML and the Data Mining Group can be found at http://www.dmg.org.
This package generates PMML for party objects, and is not a part of the PMML package for R.
This package is developed for own use (i.e., CHAID).
}
\examples{
library("CHAID")
### fit tree to subsample
set.seed(290875)
USvoteS <- USvote[sample(1:nrow(USvote), 1000),]
ctrl <- chaid_control(minsplit = 200, minprob = 0.1)
formula <- vote3 ~.
chaidUS <- chaid(formula = formula, data = USvoteS, control = ctrl)

library("pmmlParty")
pmml <- pmmlparty(chaidUS, formula, USvoteS)
pmml
}
\seealso{
\code{\link[pmml]{pmml.rpart}} \code{\link[pmml]{pmml.randomForest}}
}
